requirements.txt
requests
pendulum
pyyaml
python-dotenv
selenium
webdriver_manager
apache-airflow-providers-postgres